Retire two more sites the sanitizer stateless jobs found
`QuantilesCommon::magnitude` takes |x| through the unsigned type because, as the comment above it
says, the magnitude of the signed minimum has no signed representation. Annotate it rather than route
it through `common::negateIgnoreOverflow`: the template covers `Int8` and `Int16`, where the helper's
`auto` return integer-promotes to `int` and the narrowing back fails `-Werror`, the same way it did in
`CompressionCodecGCD`.

`ArrayJoinAction` reads `nested_offsets[index - 1]` with `index` possibly zero, which wraps before
`PaddedPODArray::operator[]` takes its `ssize_t`. This is the idiom this PR has already retired in
about 370 places; `static_cast<ssize_t>(index) - 1` matches them.

Neither is a regression, and nothing previously annotated re-reported. `Unit tests (asan_ubsan,
function_prop_fuzzer)`, which alone produced 18 sites in the first round and reaches every SQL
function, passed clean, as did `Integration tests (amd_asan_ubsan, targeted)` - so CI now confirms
independently that the 32 suppressions repaired in `f6ac2905f70b` are in force.

+++ src/Interpreters/ArrayJoinAction.cpp
@@ -235,7 +235,7 @@ void ArrayJoinResultIterator::initAnyArray()
     for (size_t row = 0; row < replicated_offsets.size(); ++row)
     {
         size_t index = indexes.getIndexAt(row);
-        accumulated += nested_offsets[index] - nested_offsets[index - 1];
+        accumulated += nested_offsets[index] - nested_offsets[static_cast<ssize_t>(index) - 1];
         replicated_offsets[row] = accumulated;
     }
 }
